requirements.txt
mcp>=1.0.0; python_version >= "3.10"
httpx>=0.27.0
python-dotenv>=1.0.0
